# project/app/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "IoT Device and Sensor API",
    "description": "API documentation for interacting with IoT sensor data, device control, and log queries.",
    "version": "1.0.0"
  },
  "host": "localhost:5002",
  "basePath": "/api/v1",
  "schemes": ["http"],
  "paths": {
    "/sensor/pulldata": {
      "get": {
        "summary": "Get sensor data",
        "description": "Retrieve sensor data based on real-time or minute-based intervals, optionally in reverse order.",
        "parameters": [
          {
            "name": "realtime",
            "in": "query",
            "required": false,
            "type": "integer",
            "enum": [0, 1],
            "description": "0 for minute-based data, 1 for real-time (second-based) data"
          },
          {
            "name": "reverse",
            "in": "query",
            "required": false,
            "type": "integer",
            "enum": [0, 1],
            "description": "0 for ascending order, 1 for descending order"
          }
        ],
        "responses": {
          "200": {
            "description": "Array of sensor data",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": { "type": "integer" },
                  "timestamp": { "type": "string", "format": "date-time" },
                  "light_level": { "type": "number" },
                  "humidity": { "type": "number" },
                  "temperature": { "type": "number" }
                }
              }
            }
          }
        }
      }
    },
    "/device/status": {
      "get": {
        "summary": "Get device status",
        "description": "Retrieve the log of device control actions such as turning devices on or off.",
        "responses": {
          "200": {
            "description": "Array of device control actions",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": { "type": "integer" },
                  "timestamp": { "type": "string", "format": "date-time" },
                  "command": { "type": "string" },
                  "status": { "type": "string" },
                  "topic": { "type": "string" }
                }
              }
            }
          }
        }
      }
    },
    "/device/control": {
      "post": {
        "summary": "Control a device",
        "description": "Send a command to control a device (e.g., turning devices on or off).",
        "parameters": [
          {
            "name": "command",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "cmd": {
                  "type": "string",
                  "enum": ["all on", "all off", "fan on", "fan off", "light on", "light off", "air on", "air off"]
                }
              }
            },
            "description": "The command to control the device."
          }
        ],
        "responses": {
          "200": {
            "description": "Device control command status",
            "schema": {
              "type": "object",
              "properties": {
                "command": { "type": "string" },
                "status": { "type": "string" }
              }
            }
          }
        }
      }
    },
    "/log/query": {
      "post": {
        "summary": "Query logs",
        "description": "Query data from the action history or sensor data based on specified parameters.",
        "parameters": [
          {
            "name": "logQuery",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "enum": ["sensorLog", "actionLog"],
                  "description": "Specify the type of log to query"
                },
                "startDate": {
                  "type": "string",
                  "format": "date",
                  "description": "Start date for the query in YYYY-MM-DD format"
                },
                "endDate": {
                  "type": "string",
                  "format": "date",
                  "description": "End date for the query in YYYY-MM-DD format"
                },
                "sort": {
                  "type": "string",
                  "enum": ["latest", "oldest"],
                  "description": "Sort order for the log data"
                },
                "numberOfRecords": {
                  "type": "integer",
                  "description": "Maximum number of records to retrieve"
                },
                "filters": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "timestamp": { "type": "string" },
                    "light_level": { "type": "string" },
                    "humidity": { "type": "string" },
                    "temperature": { "type": "string" }
                  },
                  "description": "Optional filters for specific fields in the logs"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of log entries matching the query",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": { "type": "integer" },
                  "timestamp": { "type": "string", "format": "date-time" },
                  "light_level": { "type": "number" },
                  "humidity": { "type": "number" },
                  "temperature": { "type": "number" }
                }
              }
            }
          }
        }
      }
    }
  }
}